Factor out generateSources
